moves files to more logical locations
