+++ src/bin/__mocks__/child_process.ts
