+++ src/bin/compile.ts
