+++ src/bin/config.ts
