+++ src/bin/runCommand.ts
